fix: update dependencies, switch from eslint + prettier to biome,
BREAKING CHANGE: Support for Node.js 16 is dropped, as it is End of Life.
Our code is tested against Node.js LTS release version 18 and 20.

There are a lot of changes in this commit, because of formatting changes (e.g. tabs instead of spaces).

Configured vitest to generate a junit test report, which is uploaded to CircleCI. The CI will now run biome check, which checks the code for linting and formatting issues.

Signed-off-by: Pascal Sthamer <10992664+P4sca1@users.noreply.github.com>